CLaSH.Class.BitPack

Description

Documentation

class BitPack a where

Convert to and from a BitVector

Minimal complete definition

pack, unpack

Associated Types

type BitSize a :: Nat

Number of Bits needed to represents elements of type a

Methods

pack :: a -> BitVector (BitSize a)

Convert element of type a to a BitVector

>>> pack (-5 :: Signed 6)
11_1011

unpack :: BitVector (BitSize a) -> a

Convert a BitVector to an element of type a

>>> pack (-5 :: Signed 6)
11_1011
>>> let x = pack (-5 :: Signed 6)
>>> unpack x :: Unsigned 6
59
>>> pack (59 :: Unsigned 6)
11_1011

bitCoerce :: (BitPack a, BitPack b, BitSize a ~ BitSize b) => a -> b

Coerce a value from one type to another through its bit representation.

>>> pack (-5 :: Signed 6)
11_1011
>>> bitCoerce (-5 :: Signed 6) :: Unsigned 6
59
>>> pack (59 :: Unsigned 6)
11_1011

boolToBV :: KnownNat n => Bool -> BitVector (n + 1)

Zero-extend a Boolean value to a BitVector of the appropriate size.

>>> boolToBV True :: BitVector 6
00_0001
>>> boolToBV False :: BitVector 6
00_0000