CLaSH.Class.BitVector

Documentation

class BitVector a where

Convert types from and to a vector of Bits

Associated Types

type BitSize a :: Nat

Number of Bits needed to represents elements of type a

Methods

toBV :: KnownNat (BitSize a) => a -> Vec (BitSize a) Bit

Convert element of type a to a Vec of Bits

fromBV :: KnownNat (BitSize a) => Vec (BitSize a) Bit -> a

Convert a Vec of Bits to an element of type a