CLaSH.Class.Num

Documentation

class Add a b where

Implicitly adding or subtracting values of two different (sub-)types.

Associated Types

type AResult a b

Methods

plus :: a -> b -> AResult a b

minus :: a -> b -> AResult a b

class Mult a b where

Implicitly multiplying values of two different (sub-)types.

Associated Types

type MResult a b

Methods

mult :: a -> b -> MResult a b

class Resize f where

Coerce a value to be represented by a different number of bits

Methods

resize :: (KnownNat a, KnownNat b) => f a -> f b