CLaSH.Prelude.RAM

Description

RAM primitives with a combinational read port.

RAM synchronised to the system clock

asyncRam

Arguments

:: Enum addr

=> SNat n
Size n of the RAM

-> Signal addr
Read address r

-> Signal (Maybe (addr, a))
(write address w, value to write)

-> Signal a
Value of the RAM at address r

Create a RAM with space for n elements.

NB: Initial content of the RAM is undefined

Additional helpful information:

See CLaSH.Prelude.BlockRam for more information on how to use a RAM.

asyncRamPow2

Arguments

:: KnownNat n

=> Signal (Unsigned n)
Read address r

-> Signal (Maybe (Unsigned n, a))
(write address w, value to write)

-> Signal a
Value of the RAM at address r

Create a RAM with space for 2^n elements

NB: Initial content of the RAM is undefined

Additional helpful information:

See CLaSH.Prelude.BlockRam for more information on how to use a RAM.

RAM synchronised to an arbitrary clock

asyncRam'

Arguments

:: Enum addr

=> SClock wclk
Clock to which to synchronise the write port of the RAM

-> SClock rclk
Clock to which the read address signal, r, is synchronised

-> SNat n
Size n of the RAM

-> Signal' rclk addr
Read address r

-> Signal' wclk (Maybe (addr, a))
(write address w, value to write)

-> Signal' rclk a
Value of the RAM at address r

Create a RAM with space for n elements

NB: Initial content of the RAM is undefined

Additional helpful information:

See CLaSH.Prelude.BlockRam for more information on how to use a RAM.

asyncRamPow2'

Arguments

:: KnownNat n

=> SClock wclk
Clock to which to synchronise the write port of the RAM

-> SClock rclk
Clock to which the read address signal, r, is synchronised

-> Signal' rclk (Unsigned n)
Read address r

-> Signal' wclk (Maybe (Unsigned n, a))
(write address w, value to write)

-> Signal' rclk a
Value of the RAM at address r

Create a RAM with space for 2^n elements

NB: Initial content of the RAM is undefined

Additional helpful information:

See CLaSH.Prelude.BlockRam for more information on how to use a RAM.

Internal

asyncRam#

Arguments

:: SClock wclk
Clock to which to synchronise the write port of the RAM

-> SClock rclk
Clock to which the read address signal, r, is synchronised

-> SNat n
Size n of the RAM

-> Signal' rclk Int
Read address r

-> Signal' wclk Bool
Write enable

-> Signal' wclk Int
Write address w

-> Signal' wclk a
Value to write (at address w)

-> Signal' rclk a
Value of the RAM at address r

RAM primitive