CLaSH.Prelude.ROM

Description

ROMs

Asynchronous ROM

asyncRom

Arguments

:: (KnownNat n, Enum addr)

=> Vec n a
ROM content

NB: must be a constant

-> addr
Read address rd

-> a
The value of the ROM at address rd

An asynchronous/combinational ROM with space for n elements

Additional helpful information:

See CLaSH.Sized.Fixed and CLaSH.Prelude.BlockRam for ideas on how to use ROMs and RAMs

asyncRomPow2

Arguments

:: KnownNat n

=> Vec (2 ^ n) a
ROM content

NB: must be a constant

-> Unsigned n
Read address rd

-> a
The value of the ROM at address rd

An asynchronous/combinational ROM with space for 2^n elements

Additional helpful information:

See CLaSH.Sized.Fixed and CLaSH.Prelude.BlockRam for ideas on how to use ROMs and RAMs

Synchronous ROM synchronised to the system clock

rom

Arguments

:: (KnownNat n, KnownNat m)

=> Vec n a
ROM content

NB: must be a constant

-> Signal (Unsigned m)
Read address rd

-> Signal a
The value of the ROM at address rd

A ROM with a synchronous read port, with space for n elements

NB: Read value is delayed by 1 cycle

NB: Initial output value is undefined

Additional helpful information:

See CLaSH.Sized.Fixed and CLaSH.Prelude.BlockRam for ideas on how to use ROMs and RAMs

romPow2

Arguments

:: KnownNat n

=> Vec (2 ^ n) a
ROM content

NB: must be a constant

-> Signal (Unsigned n)
Read address rd

-> Signal a
The value of the ROM at address rd

A ROM with a synchronous read port, with space for 2^n elements

NB: Read value is delayed by 1 cycle

NB: Initial output value is undefined

Additional helpful information:

See CLaSH.Sized.Fixed and CLaSH.Prelude.BlockRam for ideas on how to use ROMs and RAMs

Synchronous ROM synchronised to an arbitrary clock

rom'

Arguments

:: (KnownNat n, Enum addr)

=> SClock clk
Clock to synchronize to

-> Vec n a
ROM content

NB: must be a constant

-> Signal' clk addr
Read address rd

-> Signal' clk a
The value of the ROM at address rd from the previous clock cycle

A ROM with a synchronous read port, with space for n elements

NB: Read value is delayed by 1 cycle

NB: Initial output value is undefined

Additional helpful information:

See CLaSH.Sized.Fixed and CLaSH.Prelude.BlockRam for ideas on how to use ROMs and RAMs

romPow2'

Arguments

:: KnownNat n

=> SClock clk
Clock to synchronize to

-> Vec (2 ^ n) a
ROM content

NB: must be a constant

-> Signal' clk (Unsigned n)
Read address rd

-> Signal' clk a
The value of the ROM at address rd

A ROM with a synchronous read port, with space for 2^n elements

NB: Read value is delayed by 1 cycle

NB: Initial output value is undefined

Additional helpful information:

See CLaSH.Sized.Fixed and CLaSH.Prelude.BlockRam for ideas on how to use ROMs and RAMs

Internal

asyncRom#

Arguments

:: KnownNat n

=> Vec n a
ROM content

NB: must be a constant

-> Int
Read address rd

-> a
The value of the ROM at address rd

asyncROM primitive

rom#

Arguments

:: KnownNat n

=> SClock clk
Clock to synchronize to

-> Vec n a
ROM content

NB: must be a constant

-> Signal' clk Int
Read address rd

-> Signal' clk a
The value of the ROM at address rd from the previous clock cycle

ROM primitive