CLaSH.Signal.Bundle

Description

The Product/Signal isomorphism

Documentation

class Bundle a where

Isomorphism between a Signal of a product type (e.g. a tuple) and a product type of Signal's.

Instances of 'bundle must satisfy the following laws:

bundle . unbundle = id
unbundle . bundle = id

By default, bundle and unbundle, are defined as the identity, that is, writing:

data D = A | B

instance 'bundle D

is the same as:

data D = A | B

instance 'bundle D where
  type Unbundled' clk D = Signal' clk D
  bundle   _ s = s
  unbundle _ s = s

Associated Types

type Unbundled' (clk :: Clock) a = res | res -> clk

Methods

bundle :: Unbundled' clk a -> Signal' clk a

Example:

bundle :: (Signal' clk a, Signal' clk b) -> Signal' clk (a,b)

However:

bundle :: Signal' clk Bit -> Signal' clk Bit

bundle :: Signal' clk a -> Signal' clk a

Example:

bundle :: (Signal' clk a, Signal' clk b) -> Signal' clk (a,b)

However:

bundle :: Signal' clk Bit -> Signal' clk Bit

unbundle :: Signal' clk a -> Unbundled' clk a

Example:

unbundle :: Signal' clk (a,b) -> (Signal' clk a, Signal' clk b)

However:

unbundle :: Signal' clk Bit -> Signal' clk Bit

unbundle :: Signal' clk a -> Signal' clk a

Example:

unbundle :: Signal' clk (a,b) -> (Signal' clk a, Signal' clk b)

However:

unbundle :: Signal' clk Bit -> Signal' clk Bit