CLaSH.Signal.Explicit

Explicitly clocked synchronous signal

data CSignal clk a

A synchronized signal with elements of type a, explicitly synchronized to a clock with period clk

Clock domain crossing

newtype Clock clk

A clock with period clk

Constructors

Clock (SNat clk)

veryUnsafeSynchronizer

Arguments

:: Clock clk1
Clock of the incoming signal

-> Clock clk2
Clock of the outgoing signal

-> CSignal clk1 a

-> CSignal clk2 a

Synchronisation function that is basically a represented by a (bundle of) wire(s) in hardware. This function should only be used as part of a proper synchronisation component, such as a dual flip-flop synchronizer, or a FIFO with an asynchronous memory element:

dualFlipFlop :: Clock clk1 -> Clock clk2
             -> CSignal clk1 Bit -> CSignal clk2 Bit
dualFlipFlop clk1 clk2 = cregister clk2 L . cregister clk2 L . veryUnsafeSynchronizer clk1 clk2

Currently, this veryUnsafeSynchronizer function oversimplifies oversampling and compression: oversampling and compression rates are calculated using integer division (round towards 0).

So, given:

clk7 = Clock d7
clk6 = Clock d6
clk2 = Clock d4

Oversampling from clk6 to clk2:

>>> csample (veryUnsafeSynchronizer clk6 clk2 (cfromList [1,2,3,4,5,6,7,8,9,10]))
[1,1,1,2,2,2, 3,3,3,4,4,4, 5,5,5,6,6,...

Gives the same result as "oversampling" from clk7 to clk2:

>>> csample (veryUnsafeSynchronizer clk7 clk2 (cfromList [1,2,3,4,5,6,7,8,9,10]))
[1,1,1,2,2,2, 3,3,3,4,4,4, 5,5,5,6,6,...

That is, 2 samples at clk7 are "oversampled" to only 6 samples at clk2; meaning that there is actually a slight undersampling! A "better", but currently unimplemented, synchronisation function would give the following result:

>>> csample (veryUnsafeSynchronizer clk7 clk2 (cfromList [1,2,3,4,5,6,7,8,9,10]))
[1,1,1,2,2,2,2, 3,3,3,4,4,4,4, 5,5,5,6,6,...

Where 2 samples at clk7 are properly oversampled to 7 samples at clk2.

fromImplicit :: Signal a -> CSignal 1000 a

Implicitly clocked signals have a clock with period 1000

fromExplicit :: CSignal 1000 a -> Signal a

Implicitly clocked signals have a clock with period 1000

Basic circuit functions

csignal :: a -> CSignal t a

Create a constant CSignal from a combinational value

>>> csample (csignal 4)
[4, 4, 4, 4, ...

cregister :: Clock clk -> a -> CSignal clk a -> CSignal clk a

cregister i s delays the values in CSignal s for one cycle, and sets the value at time 0 to i

clk100 = Clock d100

>>> csampleN 3 (cregister d100 8 (fromList [1,2,3,4]))
[8,1,2]

class CPack a where

Conversion between a Signal of a product type (e.g. a tuple) and a product type of Signals

Associated Types

type CSignalP clk a

Methods

cpack :: Clock clk -> CSignalP clk a -> CSignal clk a

pack :: (Signal a, Signal b) -> Signal (a,b)

However:

pack :: Signal Bit -> Signal Bit

cunpack :: Clock clk -> CSignal clk a -> CSignalP clk a

unpack :: Signal (a,b) -> (Signal a, Signal b)

However:

unpack :: Signal Bit -> Signal Bit

Simulation functions

csimulate :: (CSignal clk1 a -> CSignal clk2 b) -> [a] -> [b]

Simulate a (CSignal clk1 a -> Signal clk2 b) function given a list of samples of type a

>>> simulate (register 8) [1, 2, 3, ...
[8, 1, 2, 3, ...

csimulateP

Arguments

:: (CPack a, CPack b)

=> Clock clk1
Clock of the incoming signal

-> Clock clk2
Clock of the outgoing signal

-> (CSignalP clk1 a -> CSignalP clk2 b)
Function to simulate

-> [a]

-> [b]

Simulate a (CSignalP clk1 a -> CSignalP clk2 b) function given a list of samples of type a

clk100 = Clock d100

>>> csimulateP clk100 clk100 (cunpack clk100 . cregister clk100 (8,8) . cpack clk100) [(1,1), (2,2), (3,3), ...
[(8,8), (1,1), (2,2), (3,3), ...

List <-> CSignal conversion

csample :: CSignal t a -> [a]

Get an infinite list of samples from a CSignal

The elements in the list correspond to the values of the CSignal at consecutive clock cycles

csample s == [s0, s1, s2, s3, ...

csampleN :: Int -> CSignal t a -> [a]

Get a list of n samples from a CSignal

The elements in the list correspond to the values of the CSignal at consecutive clock cycles

csampleN 3 s == [s0, s1, s2]

cfromList :: [a] -> CSignal t a

Create a CSignal from a list

Every element in the list will correspond to a value of the signal for one clock cycle.

NB: Simulation only!

>>> csampleN 2 (cfromList [1,2,3,4,5])
[1,2]