CLaSH.Sized.Fixed

Description

Fixed point numbers

The Num operators for the given types saturate on overflow, and use truncation as the rounding method.

BEWARE: rounding by truncation introduces a sign bias!

Truncation for positive numbers effectively results in: round towards zero

Truncation for negative numbers effectively results in: round towards -infinity

Signed fixed point numbers

type SFixed int frac = Fixed frac Signed (int + frac)

Signed Fixed-point number, with int integer bits (including sign-bit) and frac fractional bits

The Num operators for this type saturate on overflow, and use truncation as the rounding method.

*Main> maxBound :: SFixed 3 4
3.9375
*Main> minBound :: SFixed 3 4
-4.0
*Main> (1 :: SFixed 3 4) + (2 :: SFixed 3 4)
3.0
*Main> (2 :: SFixed 3 4) + (3 :: SFixed 3 4)
3.9375
*Main> (-2 :: SFixed 3 4) + (-3 :: SFixed 3 4)
-4.0
*Main> (sf d4 22 :: SFixed 3 4) * (sf d4 (-13) :: SFixed 3 4)
-1.125
*Main> (sf d4 22 :: SFixed 3 4) `mult` (sf d4 (-13) :: SFixed 3 4) :: SFixed 6 8
-1.1171875
*Main> (2 :: SFixed 3 4) `plus` (3 :: SFixed 3 4) :: SFixed 4 4
5.0
*Main> (-2 :: SFixed 3 4) `plus` (-3 :: SFixed 3 4) :: SFixed 4 4
-5.0

sf

Arguments

:: SNat frac
Position of the virtual point

-> Signed (int + frac)
The Signed integer

-> SFixed int frac

Treat a Signed integer as a Signed Fixed-point integer

*Main> sf d4 (-22 :: Signed 7)
-1.375

unSF :: SFixed int frac -> Signed (int + frac)

See the underlying representation of a Signed Fixed-point integer

Unsigned fixed point numbers

type UFixed int frac = Fixed frac Unsigned (int + frac)

Unsigned Fixed-point number, with int integer bits and frac fractional bits

The Num operators for this type saturate on overflow, and use truncation as the rounding method.

*Main> maxBound :: UFixed 3 4
7.9375
*Main> minBound :: UFixed 3 4
0.0
*Main> (1 :: UFixed 3 4) + (2 :: UFixed 3 4)
3.0
*Main> (2 :: UFixed 3 4) + (6 :: UFixed 3 4)
7.9375
*Main> (1 :: UFixed 3 4) - (3 :: UFixed 3 4)
0.0
*Main> (uf d4 22 :: UFixed 3 4) * (uf d4 (13) :: UFixed 3 4)
1.0625
*Main> (uf d4 22 :: UFixed 3 4) `mult` (uf d4 (13) :: UFixed 3 4) :: UFixed 6 8
1.1171875
*Main> (2 :: UFixed 3 4) `plus` (6 :: UFixed 3 4) :: UFixed 4 4
8.0

However, minus does not saturate to minBound on underflow:

*Main> (1 :: UFixed 3 4) `minus` (3 :: UFixed 3 4) :: UFixed 4 4
14.0

uf

Arguments

:: SNat frac
Position of the virtual point

-> Unsigned (int + frac)
The Unsigned integer

-> UFixed int frac

Treat an Unsigned integer as a Unsigned Fixed-point number

*Main> uf d4 (92 :: Unsigned 7)
5.75

unUF :: UFixed int frac -> Unsigned (int + frac)

See the underlying representation of an Unsigned Fixed-point integer

Fixed point wrapper

newtype Fixed frac rep size

Fixed-point number

Where:

frac denotes the position of the virtual point counting from the LSB

rep is the underlying representation

size is the number of bits used to represent the number

The Num operators for this type saturate on overflow, and use truncation as the rounding method.

Constructors

Fixed

Fields

unFixed :: rep size

fracShift :: KnownNat frac => Fixed frac rep size -> Int

Get the position of the virtual point of a Fixed-point number

resizeF :: forall frac1 frac2 rep size1 size2. (Bounded (rep size2), Eq (rep size1), Ord (rep size1), Num (rep size1), Bits (rep size1), Resize rep, KnownNat size2, KnownNat size1, Bits (rep size2), KnownNat frac2, KnownNat frac1, Bounded (rep size1)) => Fixed frac1 rep size1 -> Fixed frac2 rep size2

Saturating resize operation, truncates for rounding

satN2

Arguments

:: (BitVector (rep (((n + 1) + 1) + 1)), BitVector (rep (n + 1)), BitSize (rep (((n + 1) + 1) + 1)) ~ (((n + 1) + 1) + 1), BitSize (rep (n + 1)) ~ (n + 1), KnownNat (((n + 1) + 1) + 1), KnownNat (n + 1), Bounded (rep (n + 1)), Bits (rep (((n + 1) + 1) + 1)))

=> rep (((n + 1) + 1) + 1)
Number to saturate

-> rep (n + 1)
Saturated number

Resize an (N + 2)-bits number to an N-bits number, saturates to minBound or maxBound when the argument does not fit within the representations bounds of the result.

Uses cheaper saturation than resizeF, made possible by only reducing the size by 2 bit.

Proxy

asFracProxy :: Fixed frac rep size -> Proxy frac

asRepProxy :: Fixed frac rep size -> Proxy rep