CLaSH.Sized.Internal.Unsigned

Description

Datatypes

newtype Unsigned n

Arbitrary-width unsigned integer represented by n bits

Given n bits, an Unsigned n number has a range of: [0 .. 2^n-1]

NB: The Num operators perform wrap-around on overflow. If you want saturation on overflow, check out the SaturatingNum class.

>>> maxBound :: Unsigned 3
7
>>> minBound :: Unsigned 3
0
>>> read (show (maxBound :: Unsigned 3)) :: Unsigned 3
7
>>> 1 + 2 :: Unsigned 3
3
>>> 2 + 6 :: Unsigned 3
0
>>> 1 - 3 :: Unsigned 3
6
>>> 2 * 3 :: Unsigned 3
6
>>> 2 * 4 :: Unsigned 3
0
>>> (2 :: Unsigned 3) `times` (4 :: Unsigned 3) :: Unsigned 6
8
>>> (2 :: Unsigned 3) `plus` (6 :: Unsigned 3) :: Unsigned 4
8
>>> satPlus SatSymmetric 2 6 :: Unsigned 3
7
>>> satMin SatSymmetric 2 3 :: Unsigned 3
0

Constructors

U
The constructor, U, and the field, unsafeToInteger, are not synthesisable.

Fields

unsafeToInteger :: Integer

Accessors

Length information

size# :: KnownNat n => Unsigned n -> Int

Type classes

BitConvert

pack# :: Unsigned n -> BitVector n

unpack# :: BitVector n -> Unsigned n

Eq

eq# :: Unsigned n -> Unsigned n -> Bool

neq# :: Unsigned n -> Unsigned n -> Bool

Ord

lt# :: Unsigned n -> Unsigned n -> Bool

ge# :: Unsigned n -> Unsigned n -> Bool

gt# :: Unsigned n -> Unsigned n -> Bool

le# :: Unsigned n -> Unsigned n -> Bool

Enum (not synthesisable)

enumFrom# :: KnownNat n => Unsigned n -> [Unsigned n]

enumFromThen# :: KnownNat n => Unsigned n -> Unsigned n -> [Unsigned n]

enumFromTo# :: Unsigned n -> Unsigned n -> [Unsigned n]

enumFromThenTo# :: Unsigned n -> Unsigned n -> Unsigned n -> [Unsigned n]

Bounded

minBound# :: Unsigned n

maxBound# :: forall n. KnownNat n => Unsigned n

Num

(+#) :: forall n. KnownNat n => Unsigned n -> Unsigned n -> Unsigned n

(-#) :: forall n. KnownNat n => Unsigned n -> Unsigned n -> Unsigned n

(*#) :: forall n. KnownNat n => Unsigned n -> Unsigned n -> Unsigned n

negate# :: forall n. KnownNat n => Unsigned n -> Unsigned n

fromInteger# :: KnownNat n => Integer -> Unsigned n

ExtendingNum

plus# :: Unsigned m -> Unsigned n -> Unsigned (Max m n + 1)

minus# :: forall m n. (KnownNat m, KnownNat n) => Unsigned m -> Unsigned n -> Unsigned (Max m n + 1)

times# :: Unsigned m -> Unsigned n -> Unsigned (m + n)

Integral

quot# :: Unsigned n -> Unsigned n -> Unsigned n

rem# :: Unsigned n -> Unsigned n -> Unsigned n

toInteger# :: Unsigned n -> Integer

Bits

and# :: Unsigned n -> Unsigned n -> Unsigned n

or# :: Unsigned n -> Unsigned n -> Unsigned n

xor# :: Unsigned n -> Unsigned n -> Unsigned n

complement# :: KnownNat n => Unsigned n -> Unsigned n

shiftL# :: KnownNat n => Unsigned n -> Int -> Unsigned n

shiftR# :: Unsigned n -> Int -> Unsigned n

rotateL# :: KnownNat n => Unsigned n -> Int -> Unsigned n

rotateR# :: KnownNat n => Unsigned n -> Int -> Unsigned n

Resize

resize# :: forall n m. KnownNat m => Unsigned n -> Unsigned m