CLaSH.Sized.Signed

Documentation

data Signed n

Arbitrary-width signed integer represented by n bits

resizeS_wrap :: KnownNat m => Signed n -> Signed m

A resize operation that is sign-preserving on extension, but wraps on truncation.

Increasing the size of the number replicates the sign bit to the left. Truncating a number of length N to a length L just removes the leftmost N-L bits.