CLaSH.Sized.Unsigned

Documentation

data Unsigned n

Arbitrary-width unsigned integer represented by n bits

Given n bits, an Unsigned n number has a range of: [0 .. 2^n-1]

NB: The Num operators perform wrap-around on overflow. If you want saturation on overflow, check out the satN2 function in CLaSH.Sized.Fixed.