CLaSH.Class.Num

Description

Arithmetic functions for arguments and results of different precision

class ExtendingNum a b where

Adding, subtracting, and multiplying values of two different (sub-)types.

Minimal complete definition

plus, minus, times

Associated Types

type AResult a b

Type of the result of the addition or subtraction

type MResult a b

Type of the result of the multiplication

Methods

plus :: a -> b -> AResult a b

Add values of different (sub-)types, return a value of a (sub-)type that is potentially different from either argument.

minus :: a -> b -> AResult a b

Subtract values of different (sub-)types, return a value of a (sub-)type that is potentially different from either argument.

times :: a -> b -> MResult a b

Multiply values of different (sub-)types, return a value of a (sub-)type that is potentially different from either argument.

Saturating arithmetic functions

data SaturationMode

Determine how overflow and underflow are handled by the functions in SaturatingNum

Constructors

SatWrap
Wrap around on overflow and underflow

SatBound
Become maxBound on overflow, and minBound on underflow

SatZero
Become 0 on overflow and underflow

SatSymmetric
Become maxBound on overflow, and (minBound - 1) on underflow for signed numbers, and minBound for unsigned numbers.

class (Bounded a, Num a) => SaturatingNum a where

Num operators in which overflow and underflow behaviour can be specified using SaturationMode.

Minimal complete definition

satPlus, satMin, satMult

Methods

satPlus :: SaturationMode -> a -> a -> a

Addition with parametrisable over- and underflow behaviour

satMin :: SaturationMode -> a -> a -> a

Subtraction with parametrisable over- and underflow behaviour

satMult :: SaturationMode -> a -> a -> a

Multiplication with parametrisable over- and underflow behaviour

boundedPlus :: SaturatingNum a => a -> a -> a

Addition that clips to maxBound on overflow, and minBound on underflow

boundedMin :: SaturatingNum a => a -> a -> a

Subtraction that clips to maxBound on overflow, and minBound on underflow

boundedMult :: SaturatingNum a => a -> a -> a

Multiplication that clips to maxBound on overflow, and minBound on underflow