CLaSH.Tutorial

Description

A little tutorial to CλaSH

Introduction

MAC Example

Conclusion

Errors and their solutions

Type error: Couldn't match expected type ‘Signal (a,b)’ with actual type ‘(Signal a, Signal b)’:
Signals of product types and product types (to which tuples belong) of signals are isomorphic, but not (structurally) equal. Use the pack function to convert from a product type to the signal type. So if your code which gives the error looks like:

... = f a b (c,d)

add the pack function like so:

... = f a b (pack (c,d))

Product types supported by pack are:

All tuples until and including 8-tuples

The Vector type

NB: Use cpack when you are using explicitly clocked CSignals

Type error: Couldn't match expected type ‘(Signal a, Signal b)’ with actual type ‘Signal (a,b)’:
Product types (to which tuples belong) of signals and signals of product types are isomorphic, but not (structurally) equal. Use the unpack function to convert from a signal type to the product type. So if your code which gives the error looks like:

(c,d) = f a b

add the unpack function like so:

(c,d) = unpack (f a b)

Product types supported by unpack are:

All tuples until and including 8-tuples

The Vector type

NB: Use cunpack when you are using explicitly clocked CSignals

CLaSH.Normalize(94): Expr belonging to bndr: <FUNCTION> remains recursive after normalization:

If you actually wrote a recursive function, rewrite it to a non-recursive one :-)

You defined a recursively defined value, but left it polymorphic:

topEntity x y = acc
where
  acc = register 3 (x*y + acc)

The above function, works for any number-like type. This means that acc is a recursively defined polymorphic value. Adding a monomorphic type annotation makes the error go away.

topEntity :: Signal (Signed 8) -> Signal (Signed 8) -> Signal (Signed 8)
topEntity x y = acc
where
  acc = register 3 (x*y + acc)

CLaSH.Normalize.Transformations(155): InlineNonRep: <FUNCTION> already inlined 100 times in:<FUNCTION>, <TYPE>:
You left the topEntity function polymorphic or higher-order: use :t topEntity to check if the type is indeed polymorphic or higher-order. If it is, add a monomorphic type signature, and / or supply higher-order arguments.

Can't make testbench for: <LONG_VERBATIM_COMPONENT_DESCRIPTION>:

Don't worry, it's actually only a warning.

The topEntity function does not have exactly 1 argument. If your topEntity has no arguments, you're out of luck for now. If it has multiple arguments, consider bundling them in a tuple.

Haskell stuff that the CλaSH compiler does not support

Here is a list of Haskell features which the CλaSH compiler cannot synthesize to VHDL (for now):

Recursive functions

Although it seems rather bad that a compiler for a functional language does not support recursion, this bug/feature of the CλaSH compiler is amortized by the builtin knowledge of all the functions listed in CLaSH.Sized.Vector. And as you saw in this tutorial, the higher-order functions of CLaSH.Sized.Vector can cope with many of the recursive design patterns found in circuit design.

Also note that although recursive functions are not supported, recursively (tying-the-knot) defined values are supported (as long as these values do not have a function type). An example that uses recursively defined values is the following function that performs one iteration of bubble sort:

sortVL xs = vmap fst sorted <: (snd (vlast sorted))
 where
   lefts  = vhead xs :> vmap snd (vinit sorted)
   rights = vtail xs
   sorted = vzipWith compareSwapL (lazyV lefts) rights

Where we can clearly see that lefts and sorted are defined in terms of each other.

Recursive datatypes

The CλaSH compiler needs to be able to determine a bit-size for any value that will be represented in the eventual circuit. More specifically, we need to know the maximum number of bits needed to represent a value. While this is trivial for values of the elementary types, sum types, and product types, putting a fixed upper bound on recursive types is not (always) feasible. The only recursive type that is currently supported by the CλaSH compiler is the Vector type, for which the compiler has hard-coded knowledge.

GADT pattern matching

While pattern matching for regular ADTs is supported, pattern matching for GADTs is not. The Vector type, which is also a GADT, is no exception! You can use the extraction and indexing functions of CLaSH.Sized.Vector to get access to individual ranges / elements of a Vector.